+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["schemd"]
+members = ["schemd/cli"]
 
 [package]
 name = "cargo-deny"
