--- config-spec/schemd/lib.js
@@ -1,6 +1,0 @@
-import { createTypeSpecLibrary } from "@typespec/compiler";
-
-export const $lib = createTypeSpecLibrary({
-    name: "schemd",
-    diagnostics: {},
-});
