+++ cspell.yml
@@ -1,3 +1,4 @@
 words:
   - schemd
   - rekeyable
+  - typespec
