+++ schemd/cli/src/cli.rs
