+++ schemd/cli/src/cli/codegen/json_schema.rs
