+++ schemd/cli/src/entrypoint.rs
