+++ schemd/cli/src/error.rs
