+++ schemd/cli/src/lib.rs
