+++ schemd/cli/src/main.rs
