+++ schemd/tsp/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./decorators/index.js";
+export { $lib } from "./lib.js";
+export const namespace = "Schemd";
