+++ schemd/tsp/src/lib.ts
@@ -1,0 +1,6 @@
+import * as tsp from "@typespec/compiler";
+
+export const $lib = tsp.createTypeSpecLibrary({
+    name: "@schemd/tsp",
+    diagnostics: {},
+});
